% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-ard.R
\name{complement}
\alias{complement}
\title{Generate data with specified correlation to a vector}
\usage{
complement(y, corr = 0.5)
}
\arguments{
\item{y}{Vector to correlate with.}

\item{corr}{Target correlation value (between -1 and 1).}
}
\value{
A vector with the specified correlation to \code{y}.
}
\description{
Generate data with specified correlation to a vector
}
\examples{
y <- rnorm(100)
x <- complement(y, corr = 0.8)
cor(x, y)
}
