% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-ard.R
\name{generate_traits}
\alias{generate_traits}
\title{Generate traits correlated with latent positions}
\usage{
generate_traits(U, corr, num_traits)
}
\arguments{
\item{U}{A matrix of latent positions.}

\item{corr}{A numeric vector (or scalar) of correlations.}

\item{num_traits}{The number of traits to generate.}
}
\value{
A matrix of traits.
}
\description{
Generate traits correlated with latent positions
}
\keyword{internal}
