% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-ard.R
\name{left_padded_sequence}
\alias{left_padded_sequence}
\title{Left-pad an integer sequence with zeros}
\usage{
left_padded_sequence(x)
}
\arguments{
\item{x}{A vector of integers.}
}
\value{
A character vector of the same length as \code{x}, left-padded with zeros.
}
\description{
Left-pad an integer sequence with zeros
}
\keyword{internal}
