% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{sample_graph}
\alias{sample_graph}
\title{Sample a graph from an estimated spectral ARD model}
\usage{
sample_graph(object, allow_self_loops = FALSE, poisson_edges = FALSE, ...)
}
\arguments{
\item{object}{A \code{SpectralArdEstimate} object.}

\item{allow_self_loops}{Logical. Should self-loops be allowed? Defaults to \code{FALSE}.}

\item{poisson_edges}{Logical. Should the number of edges between nodes be
sampled from a Poisson distribution? If \code{FALSE} (the default), edges are
sampled from a Bernoulli distribution.}

\item{...}{Currently ignored.}
}
\value{
An \code{igraph} object.
}
\description{
Sample a graph from an estimated spectral ARD model
}
\examples{

set.seed(42)

# 1. Simulate ARD data
sim <- simulate_ard_data(
  n = 100, k = 2, corr = 0.7,
  num_good_traits = 2, num_bad_traits = 2
)

# 2. Estimate the network model
estimate <- estimate_spectrum(sim$ard, sim$traits)
estimate

# 3. Sample a new graph from the estimate
g <- sample_graph(estimate)
g
}
