% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimate_spectrum}
\alias{estimate_spectrum}
\title{Estimate a low-rank network model from Aggregated Relational Data (ARD)}
\usage{
estimate_spectrum(ard, traits)
}
\arguments{
\item{ard}{A \code{matrix} containing Aggregated Relational Data. Each row of \code{ard}
corresponds to one node in the network, and each column represents an
aggregated trait. The order of the nodes (rows) and aggregated traits (columns)
should match \code{traits}. Missing data is not allowed.}

\item{traits}{A \code{matrix} containing node-level trait information. Each row of \code{traits}
corresponds to one node in the network, and each column represents a particular trait.
The order of the nodes (rows) and aggregated traits (columns) should match \code{ard}.
Missing data is not allowed.}
}
\value{
A \code{SpectralArdEstimate} object.
}
\description{
Estimate a low-rank network model from Aggregated Relational Data (ARD)
}
\examples{

set.seed(42)

# 1. Simulate ARD data
sim <- simulate_ard_data(
  n = 100, k = 2, corr = 0.7,
  num_good_traits = 2, num_bad_traits = 2
)

# 2. Estimate the network model
estimate <- estimate_spectrum(sim$ard, sim$traits)
estimate

# 3. Sample a new graph from the estimate
g <- sample_graph(estimate)
g
}
